Refactored Profile
Moved all profile features to Profile class and removed button for account management.
